install virtualbox-guest* packages from sid
At the time I built the image from testing, packages
virtualbox-guest-dkms and virtualbox-guest-x11 were
not available in testing repositories.

This patch allows to download the packages from sid
repository, but it won't add the sid repository to
the installed system. So these two packages will be
seen as obsolete by aptitude. Not a problem anyway.
I'll see when Stretch is released.