add oxygen-icon-theme
This is a recommended package of lxqt-common. live-build is
configured not to install recommended packages, in order to
limit the size of the image. But if this package is not
installed no application icon will appear on the menu by default
(you must manually change it in the LxQT settings).

Add also aptitude, best command-line tool to manage packages.